change the share CTA in accordance with #2510 (#2584)
* change the share CTA in accordance with #2510
