+++ source/js/buyers-guide/components/primary-nav/primary-nav.scss
@@ -120,28 +120,6 @@
   }
 }
 
-.social-button {
-  display: block;
-  width: 20px;
-  height: 20px;
-  margin: 0 12px;
-  background-size: contain;
-  background-repeat: no-repeat;
-  background-position: center;
-}
-
-.social-button-fb {
-  background-image: url("/_images/glyphs/fb.svg");
-}
-
-.social-button-twitter {
-  background-image: url("/_images/glyphs/twitter.svg");
-}
-
-.social-button-email {
-  background-image: url("/_images/glyphs/email.svg");
-}
-
 .btn-donate {
   background: #ff4f5e;
   border-radius: 14px;
