+++ source/js/buyers-guide/components/social-share/social-share.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactGA from '../../../react-ga-proxy';
+import copyToClipboard from '../../copy-to-clipboard.js';
 
 const SocialShareLink = (props) => {
-  let classes = `social-button`;
+  let classes = `social-icon`;
   let srLabel = ``;
   let link = `PrivacyNotIncluded.org`;
   let shareText = `I think this ${props.productName} is ${props.creepType.toUpperCase()}. What do you think? Check out the Creep-O-Meter over on @mozilla’s ${link} buyer’s guide.`;
@@ -34,19 +35,40 @@ const SocialShareLink = (props) => {
     link = `mailto:?&body=${encodeURIComponent(shareText)}`;
   }
 
+  if (props.type === `link`) {
+    classes += ` social-button-link`;
+    srLabel = `Link`;
+    shareEvent.label += `using a link`;
+    link = `#`;
+  }
+
   let trackShareAction = () => {
     ReactGA.event(shareEvent);
   };
 
-  return <a target="_blank" className={classes} href={link} onClick={trackShareAction}><span class="sr-only">{srLabel}</span></a>;
+  // We actually need slightly different behaviour
+  // for the "copy link" functionality:
+  if (props.type === `link`) {
+    let _trackShareAction = trackShareAction;
+
+    trackShareAction = (evt) => {
+      evt.preventDefault();
+      copyToClipboard(evt.target, window.location.href);
+      evt.target.innerHTML = evt.target.innerHTML.replace(srLabel, `Copied`);
+      _trackShareAction();
+    };
+  }
+
+  return <a target="_blank" className="pni-s-link" href={link} onClick={trackShareAction}><span className={classes}/> {srLabel}</a>;
 };
 
 const SocialShare = (props) => {
   return (
-    <div class="social d-flex justify-content-center mt-3">
+    <div class="social pni-share-buttons d-flex justify-content-center flex-wrap flex-md-nowrap mt-3">
       <SocialShareLink type="facebook" {...props} />
       <SocialShareLink type="twitter" {...props} />
       <SocialShareLink type="email" {...props} />
+      <SocialShareLink type="link" {...props} />
     </div>
   );
 };
