+++ source/sass/buyers-guide/includes/social.scss
@@ -1,0 +1,29 @@
+.social-button {
+  display: block;
+  width: 20px;
+  height: 20px;
+  margin: 0 12px;
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-position: center;
+
+  @include hover-focus-active {
+    text-decoration: none;
+  }
+}
+
+.social-button-fb {
+  background-image: url("/_images/glyphs/fb.svg");
+}
+
+.social-button-twitter {
+  background-image: url("/_images/glyphs/twitter.svg");
+}
+
+.social-button-email {
+  background-image: url("/_images/glyphs/email.svg");
+}
+
+.social-button-link {
+  background-image: url("/_images/glyphs/link.svg");
+}
